ShaderProgram merged and little bit refactored into this project guidelines
